Redact photo/media picker exif more often
If the application requesting the photo/media picker is not the same
as the picker provider itself, always redact sensitive exif data such as
GPS location fields.

There have been some notable exceptions to exif redaction via content
picker. Previously, redaction would never occur under if:
- the content provider is a file manager, or
- the content provider is the system gallery and has access to the file.

This prevented anything from being redacted if the system file picker or
gallery were used to the select the file.

As a quick proof of concept prior to implementing better checks, this
change simply insists on redaction when the calling application is not
the picker itself.

Issue: calyxos#2468
Change-Id: Ic6bb59a32fe9441ecc8488c97c80e1ecbb51d5a1