Redact exif when actual caller is not provider
If the application requesting the picker is not the same as the picker
itself, always redact.

There are many exceptions to exif redaction via content picker.
For example, redaction never occurs under the following conditions:
- The content provider is a file manager.
- The content provider is the system gallery and has access to the file.

This prevents anything from being redacted if the system file picker or
gallery are used to the select the file.

As a quick proof of concept prior to implementing better checks, simply
insist on redaction if the calling application is not the picker itself.

Issue: calyxos#2468
Change-Id: Ic6bb59a32fe9441ecc8488c97c80e1ecbb51d5a1